moved veHoneyModal files to one folder
